application/requirements.txt
# LIB DEPENDENCIES
requests~=2.32.3
flask~=3.1.0
aiohttp~=3.11.18
pywin32; platform_system == "Windows"

# APP DEPENDENCIES
pystray
pillow
keyring